feat(submission): add middleware
* user cannot submit submission pass assignment's due date

* user cannot submit more than one submission to an assignment

* user cannot modify others submission

* user cannot modify submission pass assignment's due date

+++ server/app/Http/Controllers/SubmissionController.php
@@ -2,15 +2,40 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Resources\AssignmentResource;
 use App\Http\Resources\SubmissionResource;
 use App\Models\Assignment;
 use App\Models\Course;
 use App\Models\Submission;
+use Closure;
 use Illuminate\Http\Request;
+use Illuminate\Routing\Controllers\HasMiddleware;
+use Illuminate\Routing\Controllers\Middleware;
 
-class SubmissionController extends Controller
+class SubmissionController extends Controller implements HasMiddleware
 {
+
+    public static function middleware(): array
+    {
+        return [
+            new Middleware('auth:api'),
+            // new Middleware('role:teacher', only: ['index', 'show']),
+            // new Middleware('role:student', except: ['index', 'show']),
+
+            // function (Request $request, Closure $next, string $course) {
+            //     $course = Course::where('slug', $course)->firstOrFail();
+
+            //     if (
+            //         !$request->user()->hasPurchased($course) ||
+            //         $course->teacher === $request->user()->id
+            //     ) {
+            //         return response()->json(['error' => 'Unauthorized'], 403);
+            //     }
+
+            //     return $next($request);
+            // }
+        ];
+    }
+
     /**
      * Display a listing of the resource.
      */
@@ -40,9 +65,14 @@ class SubmissionController extends Controller
         $assignment = Assignment::findOrFail($assignment);
 
         //check whether user has already submitted the submission
-        // if ($request->user()->hasSubmitted($assignment)) {
-        //     return response()->json(['message' => 'User has already made submission to this assignment'], 403);
-        // }
+        if ($assignment->submissions()->where('user_id', $request->user()->id)->exists()) {
+            return response()->json(['message' => 'User has already made submission to this assignment'], 403);
+        }
+
+        // check whether submission has passed assignment due date
+        if ($assignment->due_date >= now()) {
+            return response()->json(['message' => 'Submitting submission not allowed exceeding assignment\'s due date'], 406);
+        }
 
         $validated = $request->validate([
             'content' => 'present|nullable|string'
@@ -89,6 +119,16 @@ class SubmissionController extends Controller
         $course = Course::where('slug', $course)->firstOrFail();
         $submission = Submission::findOrFail($submission);
 
+        // check whether the submission belong to currently authenticated user
+        if ($submission->user_id !== $request->user()->id) {
+            return response()->json(['message' => 'Unauthorized'], 403);
+        }
+
+        // check whether submission has passed assignment due date
+        if ($submission->assignment->due_date >= now()) {
+            return response()->json(['message' => 'Submission changes not allowed exceeding assignment\'s due date'], 406);
+        }
+
         $validated = $request->validate([
             'content' => 'present|nullable|string'
         ]);
@@ -104,6 +144,7 @@ class SubmissionController extends Controller
      * Remove the specified resource from storage.
      */
     public function destroy(
+        Request $request,
         string $course,
         string $chapter,
         string $subchapter,
@@ -112,6 +153,16 @@ class SubmissionController extends Controller
     ) {
         $course = Course::where('slug', $course)->firstOrFail();
         $submission = Submission::findOrFail($submission);
+
+        // check whether the submission belong to currently authenticated user
+        if ($submission->user_id !== $request->user()->id) {
+            return response()->json(['message' => 'Unauthorized'], 403);
+        }
+
+        // check whether submission has passed assignment due date
+        if ($submission->assignment->due_date >= now()) {
+            return response()->json(['message' => 'Submission changes not allowed exceeding assignment\'s due date'], 406);
+        }
 
         $submission->delete();
